Roba per API, principalmente gruppi. Da provare.

+++ app/src/controllers/groupController.js
@@ -1,15 +1,22 @@
 const Group = require('../models/groupModel');
 const UserGroupRelation = require('../models/userGroupModel');
 const sequelize = require('../utils/dbConnect');
+const getTrueFields = require('../utils/getTrueFields');
+const sendError = require('../utils/sendError');
 
 const getGroups = async (req, res) => {
-    const groups = await Group.findAll({
-        attributes: getTrueFields(req.query)
-    });
-    if (groups)
-        res.status(200).send(groups);
-    else
-        res.status(404).send("Nessun gruppo trovato");
+    try {
+        const groups = await Group.findAll({
+            attributes: getTrueFields(req.query) || ['id', 'name', 'num_members', 'is_private']
+        });
+        if (groups)
+            res.status(200).send(groups);
+        else
+            res.status(404).send("Nessun gruppo trovato");
+    } catch (error) {
+        sendError(error, res);
+        return;
+    }
 }
 
 const getGroup = async (req, res) => {
@@ -26,23 +33,24 @@ const getGroup = async (req, res) => {
 }
 
 const createGroup = async (req, res) => {
-    const { name, num_members } = req.body;
-    if (await Group.findOne({
-        where: {
-            name: name
-        }
-    })) {
-        res.status(409).send("Nome già in uso");
-        return;
-    }
     try {
+        const { name, num_members } = req.body;
+        if (await Group.findOne({
+            where: {
+                name: name
+            }
+        })) {
+            res.status(409).send("Nome già in uso");
+            return;
+        }
         await Group.create({
             name: name,
             num_members: num_members || 0,
             is_private: false
         });
     } catch (error) {
-        res.status(501).send("Errore durante la creazione del gruppo");
+        sendError(error, res);
+        return;
     }
     res.status(200).send("Gruppo creato con successo");
 }
@@ -60,20 +68,35 @@ const updateGroup = async (req, res) => {
             }
         });
     } catch (error) {
-        res.status(501).send("Errore durante l'aggiornamento del gruppo");
+        sendError(error, res);
+        return;
     }
     res.status(200).send("Gruppo aggiornato con successo");
 }
 
 const deleteGroup = async (req, res) => {
     try {
-        await Group.destroy({
-            where: {
-                id: req.params.id
-            }
-        });
+        if (await Group.findOne({where: {id: req.params.id, is_private: false}})) {
+            const transaction = await sequelize.transaction();
+            await UserGroupRelation.destroy({
+                where: {
+                    gid: req.params.id
+                }
+            }, { transaction });
+            await Group.destroy({
+                where: {
+                    id: req.params.id
+                }
+            }, { transaction });
+            await transaction.commit();
+        } else {
+            res.status(403).send("Non puoi eliminare un gruppo privato");
+            return;
+        }
     } catch (error) {
-        res.status(501).send("Errore durante la cancellazione del gruppo");
+        await transaction.rollback();
+        sendError(error, res);
+        return
     }
     res.status(200).send("Gruppo cancellato con successo");
 }
@@ -82,15 +105,7 @@ const addUserToGroup = async (req, res) => {
     const { uid } = req.body;
     const gid = req.params.id;
     const transaction = await sequelize.transaction();
-    if (await UserGroupRelation.findOne({
-        where: {
-            uid: uid,
-            gid: gid
-        }
-    })) {
-        res.status(409).send("Utente già presente nel gruppo");
-        return;
-    } else if (await Group.findOne({
+    if(await Group.findOne({
         where: {
             id: gid,
             is_private: true
@@ -113,9 +128,43 @@ const addUserToGroup = async (req, res) => {
     } catch (error) {
         console.log(error)
         await transaction.rollback();
-        res.status(501).send("Errore durante l'aggiunta del membro");
+        sendError(error, res);
     }
     res.status(200).send("Membro aggiunto con successo");
+}
+
+const removeUserFromGroup = async (req, res) => {
+    const { uid } = req.body;
+    const gid = req.params.id;
+    const transaction = await sequelize.transaction();
+    if (await Group.findOne({
+        where: {
+            id: gid,
+            is_private: true
+        }
+    })) {
+        res.status(403).send("Gruppo privato");
+        return;
+    }
+    try {
+        await Group.decrement('num_members', {
+            by: 1,
+            where: {
+                id: gid
+            }
+        }, { transaction });
+        await UserGroupRelation.destroy({
+            where: {
+                uid: uid,
+                gid: gid
+            }
+        }, { transaction });
+        await transaction.commit();
+    } catch (error) {
+        await transaction.rollback();
+        res.status(501).send("Errore durante la rimozione del membro");
+    }
+    res.status(200).send("Membro rimosso con successo");
 }
 
 module.exports = {
@@ -124,5 +173,6 @@ module.exports = {
     createGroup,
     updateGroup,
     deleteGroup,
-    addUserToGroup
+    addUserToGroup,
+    removeUserFromGroup
 }
